net/mlx5: Add support for fw live patch event
Firmware live patch event notifies the driver that the firmware was just
updated using live patch. In such case the driver should not reload or
re-initiate entities, part to updating the firmware version and
re-initiate the firmware tracer which can be updated by live patch with
new strings database to help debugging an issue.

Issue: 1959843
Change-Id: I4406a6b0229980f657c6507bb146180304a7581f
Signed-off-by: Moshe Shemesh <moshe@mellanox.com>