RDMA/core: Drop destroy reason argument from uobjects
The initial idea was that destroy reason supplied during the destroying
of the uobjects was intended to be forwarded to the drivers. However latest
ib_core simplifications (like abort scheme, destroy return error handling,
e.t.c.) made this parameter obsolete, so clean the code to get rid of the
rdma_remove_reason in ->destroy_object() flows.

Signed-off-by: Leon Romanovsky <leonro@nvidia.com>